Use dict key accessing instead of `getattr`
We're accessing a column in a pandas dataframe and using attribute
access for this is not necessary. Since we have the column name as a
string here, using dictionary lookup syntax instead is not only the
recommended, but also the more straightforward and readable solution in
this case.
And as it turns out, using `getattr` didn't even behave correctly.