Break vs. continue: small change, huge impact
Obviously, when the cluster contains no bus for the current grouping, we
can stop handling that grouping, but we should not skip all other
groups. That's why we have to use a `continue` here instead of a
`break`.
Hopefully not only the first, but also the last time that I mixed those
two up.