Improve removal of unwanted buses
The partial network contains some buses which actually don't belong to
the subnetwork represented by the cluster. It took me a while to figure
out how to detect those buses in order to remove them from the set of
buses to which solutions are distributed, but it looks like these buses
are the external buses that have a direct connection into the cluster.
That's why removing them can be done via checking for `self.idx_prefix`
as that is what's prepended to the names of external buses.
Also, switching from `iterrows` to `itertuples` should be faster
(according to pandas' documentation) and it makes it easier to see what
is being checkt. (In this case: the name of the `bus`.)