Add a few debug assertions
The first one is there, because I stumble upon a bug which resulted in
no buses corresponding to the current cluster bus found in the subnetwork.
Obviously this shouldn't be the case.
The other two assertions check that the `storage_extendable` flag
behaves the way I'm assuming it should behave. If those assumptions are
wrong, code needs to be changed.