Add sanity check for cumulative `'p'` values
When summing up the `'p'` values for specific buses over all timesteps,
the values should be identical, module small rounding errors, for the
clustered network and the original network after uniform disaggregation.
This commit adds output for the difference of these cumulative values,
so that one can immediately spot when the numbers are too far off.