Switch more places to using `.loc`
Using `.loc` instead of standard dictionary access should be faster in
most times, according to pandas' documentation. But more importantly it
makes it more explicit whether the key is used as a column or as a row
index and it also makes the shape of the result immediately apparent.
Last but not least, it's now easier to spot commonalities between the
`if` and the `else` branch.