Refactor time dependency test
Using an appropriately called function better conveys what is done.
Also, it means that we are no longer using the local variable for
multiple different purposes, which allows us to remove the comment, as
it won't be used later. Figuring out over which time series keys to
iterate will be done differently later in subsequent commits.