Squelch last `SettingWithCopyWarning`
It's weird that the warning even gets fired, as using `.loc` should mean
there is no copying happening. But I guess the variable itself is a
partial copy from some other dataframe. As the code is still working as
intended, using pandas' built in way of squelching false positives is
the way to go.