Factor kmeans execution out into generic function
Factoring this part out is done, so that the same can be done with other
parts of the program. The order in which these factored out parts are
then executed can be controlled via the `'ordered'` entry in the `args`
dictionary. Since the call structure for the factored out functions has
to be the same for all of them, it has to be generic enough to
accommodate most use cases.
The hard part here was figuring out how to be able to specify what
arguments to pass to the functions, as these arguments might consist of
local variables as well as entries from the global `args` dictionary.
Specifically figuring out that I have to do save `locals()` because
dictionary comprehensions have their own local scope, took me a while.
This is the reason that the `variables = locals()` and `del variables`
lines are necessary. The second one isn't specifically necessary, but it
feels safer to not have an alias for `locals()` lingering around for too
long.