Distinguish clustered and disaggregated networks
The disaggregated network is mainly there to pass it on to eDisGo, while
eTraGo wants to keep on working on the clustered network. Previously,
disaggregation overwrote the clustered network with the disaggregated one.
Now, after clustering the clustered network becomes the new default
'network' and the original, unclustered network gets overwritten with
the disaggregated one later. If no disaggregation is performed, the
original network gets thrown away after clustering.