Shorten timing code
I think it's still readable enough and not having so many intermediate
variables cleans up stuff a bit. Also, the comment is lying a bit
because there is no `z`. The variable is called `c`, which is a
testament to too many intermediate variables sometimes confuse more than
they help.