Adjust `'q'` disaggregation to `allocation`
The `'q'` timeseries gets distributed to all buses on a cluster
according to a certain strategy via `distribute_q` from the
`etrago.tools.utilities` module. Depending on which `allocation`
strategy is chosen, the factors according to which `'q'` is distributed
from a cluster down to it's partial network have to be adjusted as well.
This patch aligns the choice of distribution factors with the
`allocation` strategy chosen.