Index pandas objects using boolean values directly
The changes squelch the same `flake8` complaints as the previous commit,
i.e. the one with id 21af8dc834ff655eec1c5aa17b1d50d938e660a2, but they
are made in a separate commit because of potential bad interactions with
pandas' indexing. Pandas overloads `==` and indexing in special ways, so
if `df1` and `df2` are `DataFrame`s, `df1[df2["column"] == True]` and
`df1[df2["column"]]` might mean different things, although they
shouldn't.
So in order to be able to roll back these changes easily, they are made
in a separate commit.