Use `if not path` instead of `if pass == False`
Again, squelch a `flake8` complaint. While similar to commit 90a7380
this situation is special, because the new code has additional benefits.
Using `not path` also catches the empty string and `None`, which both
would have lead to the functions full execution since they are not equal
to `False`. The former would probably have silently polluted the working
directory while the latter would have generated an error.
Last but not least: testing for `False` is weird since that is not even
a legal value according to the (old) docstring.