Squelch `flake8` warning about a bare except
This one's important, because the bare except is actually hiding a bug.
Even if there's no SQL error, the result of

  ```python
  pd.read_sql(sql, engine)`
  ```

is a `DataFrame` with a single row and column called `"gas_parameters"`
containing a dictionary. So the

  ```python
  marginal_cost = df["marginal_cost"]
  ```

line will always raise a `KeyError`. But this error is swallowed by bare
except, so one would never notice that the default for `marginal_cost`
is always used.
The correct way would probably be something like
`df["gas_parameters"]["marginal_cost"]` but we can't be sure until this
can be tested with live data.