Add "noxfile.py" with build checks
Most importantly, the checks include running `black`, `flake8` and
`isort` on files known to be clean. So when fixing a file, you can now
add it to the list of `cleaned` files in the `check` function in
"noxfile.py". Then the workflow will complain, if changes to the files
trigger errors from those tools.