Fix style "violations"
Apparently `black` released version [`23.1.0`][0] which incorporates
some new style fixes. Most importantly it started to

> "Remove trailing newlines after code block open (#3035) (22.6.0)"

which broke the CI workflow, since the checked files weren't formatted
with the most up to date version of `black`.
So the fix is easy: run the most up to date version of `black` on the
files.

[0]: https://black.readthedocs.io/en/stable/change_log.html#id1