Check packaging on multiple Python versions
The other checks should be version agnostic, so they can just use the
default version of Python 3 for the platform under test. But packaging
needs to be tested for every version so that we're sure that we support
all versions we claim to support. For this reason, explicitly build
and check the package on all Python versions we want to support and also
do so on the default version of Python 3 for the platform under test, in
case there's a new Python version out which we aren't aware of. This
catches potential future problems early.