Bulk insert disaggregated columns
Instead of using a for loop to append them one at a time. Doing so
triggered a few warnings about the `DataFrame` being fragmented. The new
code should prevent that, while still inserting the same values.
Additionally, no duplicate columns are created. Instead the data in
already existing columns is overwritten, which is a plus.