Adjust `query` for the special case of `"links"`
For links, the `query` contains references to `bus1`, which has a
different meaning depending on whether we are dealing with a cluster or
the partial network the cluster represents. For clusters, the query is
correct, because the group was generated by looking up the `bus1` values
of links attached to the cluster. But to find the corresponding links in
the partial network, one has to look for those links whose `bus1` value
matches the index of one of the buses whose `"cluster"` entry is equal
to the group's `bus1` value.
Or in other words, since the group's `bus1` value identifies a cluster,
one has to

  - first find all buses corresponding to that cluster,
  - then find all links whose `bus1` value points to one of these buses.