Filter links by `bus0` instead of `bus1`
This fixes a bug in the original `p0`, `p1` disaggregation
implementation for `links` in commit 988d19cb9. For components having a
`bus0` and a `bus1` instead of just a `bus`, `bus0` is taken as
corresponding to `bus`. So filtering out these components should be done
by looking at `bus0`, not `bus1`.