Move `bus == cluster` filtering out of the loop
Nothing in the filter uses the loop variable, so the filtering can be
moved out of the loop. Also, since `cl_buses` is only used to determine
the value of `clb` for the duration of the loop and nowhere else,
filtering can be done directly when assigning `cl_buses` and `clb` can
be determined directly by querying the filtered DataFrame.
Filtering before determining the groups massively speeds up the
disaggregation.