Use values in cluster columns to define groups
This is the correct way anyway and fixes a bug, since the values stored
in each group (should) uniquely identify a cluster. The bug just never
surfaced, as these values are usually also found in the partial network.
This isn't the case for the values of `bus`, `bus0` and `bus1` though,
because these point to other cluster or partial network buses, so they
differ depending on whether the attribute is on a cluster or a partial
network component, so the bug was finally caught.