Add links to "clustered - disaggregated" checks
This code prints the differences of the sum of all values of a certain
attribute or timeseries in the clustered network to the sum of the
corresponding disaggregated values in the original network.
Usually these should be close to zero. For `links`'s "p0" and "p1"
timeseries these aren't close to zero, but that's OK as these include
values from the CH4 links, which aren't disaggregated.
But it's probably still a good idea to have those printed in order to be
able to check whether these differences are in the correct ballpark.
Maybe the check can later be extended to be filtered so that they are
close to zero again.