Only use p_max_pu timeseries data for generators that have this data
PyPSA's networkclustering now deletes timeseries that only inlcude the default value.
When running smaller example calculations eith only a few snapshots, it could
happen, that the p_max_pu time series of some wind or solar generator is always 1.0
To avoid key errors in these cases, it was required to not use the carrier to
decide if a generator has a p_max_pu time series but directly look at the existing columns.